requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
autopep8==2.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.4
cryptography==41.0.1
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.54.0
django-autoslug==1.9.9
django-cities-light==3.9.2
django-filter==22.1
django-guardian==2.4.0
django-jazzmin==2.6.0
django-phonenumber-field==7.1.0
django-recaptcha==3.0.0
exceptiongroup==1.1.2
fontawesomefree==6.4.0
holidays==0.25
html5lib==1.1
idna==3.4
iniconfig==2.0.0
korean-lunar-calendar==0.3.1
lxml==4.9.3
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.1
phonenumbers==8.13.15
Pillow==9.5.0
pluggy==1.2.0
progressbar2==4.2.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyHanko==0.19.0
pyhanko-certvalidator==0.23.0
PyJWT==2.7.0
pypdf==3.12.1
pypng==0.20220715.0
pytest==7.4.0
pytest-django==4.5.2
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-utils==3.7.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
typing_extensions==4.6.3
tzlocal==5.0.1
Unidecode==1.3.6
uritools==4.0.1
urllib3==2.0.3
webencodings==0.5.1
xhtml2pdf==0.2.11
